include binary shared libs that are used only for linking against
In order to build the native parts, they need to be linked against openssl
and SQLCipher.  For deployments, these libraries are provided elsewhere, so
include binaries in git that are good enough for linking against.

libcrypto.so came from android-15 4.0.3 emulators from the Android SDK.
libcrypto.so is included in Android itself, and automatically loaded before
any Android app is started.  Therefore, we just need to link against it, it
should never be shipped with the app!

libsqlcipher_android.so came from SQLCipher+for+Android+v3.1.0.zip SHA-256:
9568592f14d8b28d586b742cebb4e032f1f42888a67cdb4685b217d1285a1dbd