include SDK and NDK info in the jar manifest
This is useful for reproducing the build.  Java automatically adds its own
version info to the manifest, so it will also be useful to include the
Android build tools as well.  Additionally, the manifest will only ever
exactly match if both builds use the _exact_ same Java version (i.e.
1.7.0_65-b32) so this does not break the determinism of the manifest since
it is already broken.