move setContainerPath() to JNI and handle all pre-mount checks there
Instead of having the code for pre-checking the container file in both Java
and JNI/C++, this merges it all into the JNI side.  This is also a step
towards changing the core logic of mounting so that it will fail if the
container file does not exist or is not read/write.  Previously, it would
just create the file at the path given.  This makes it hard to detect if
the path to the container file is wrong, like if the app was moved to the
SD card.

refs #4531 https://dev.guardianproject.info/issues/4531