use one static blob for generating Exception messages
Since there can only be a single Exception, there only can be a single
Exception message. So there does not need to be multiple copies of msg.
This also just simplifies the code a bit, and provides a constant to make
sure that the snprintfs are doing the right thing.  Yes, I'm lazy about C
memory management.