createNewContainer methods for creating a new IOCipher file system
This changes the VirtualFileSystem paradigm to make mount() only operate on
existing files, then introduces createNewContainer() as a dedicated method
for createing new VFS container files.

Before, VirtualFileSystem.mount() would create the container file if it did
not exist.  But this makes it quite difficult to troubleshoot problems with
the access to that file, like with:
refs #4531 https://dev.guardianproject.info/issues/4531

unmount() can no longer zero out the dbFileName because now, the user might
want to setContainerPath() once, then run mount() and unmount() without
setting the container path again.