add method to finalize sqlfs on extra threads
When a new thread tries to access the volume, it will use the cached password
and path and automatically allocate a sqlfs_t.  When that thread ends, it'll
automatically call sqlfs_t_finalize() to cleanup.  But often times threads are
not exited, but instead reused.  Like with Android's AsyncTask, it reuses a
pool of threads, so the thread is never exited.  Threading also works like
that on iOS. GCD (Grand Central Dispatch) uses a thread pool.

Using libsqlfs in a thread pool will never dealloc the sqlfs_t for each
thread. The problem is mostly about instance counting and memory deallocation,
as well as closing the connection to sqlite.  What we need is a way with
pthread to automatically run a function when a thread is reused rather than
exited.

This provides a manual workaround: call sqlfs_detach_thread() when the
thread has completed the task at hand and no longer needs access to the
IOCipher container.